Update README documentation and switch to Markdown.
The readme is updated to include a thorough example with inline documentation
and commenting. Example includes actual JSON and YAML output.

The format of the readme is changed from RST to Markdown. This is because github
has built in Markdown support and since this project is still experimental and
does not have Sphinx documentation, the simple and support Markdown is preferred.